Enable select button in Graphslice.tsx
